pyproject.toml
[project]
name = "guest-post-link-checker"
version = "0.1.0"
description = ""
authors = [
    {name = "MR11Robot",email = "taha.youssef.fares@gmail.com"}
]

license = {text = "MIT"}
requires-python = ">=3.14"
dependencies = [
    "schedule (>=1.2.2,<2.0.0)",
    "flask (>=3.1.3,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "undetected-chromedriver (>=3.5.5,<4.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "fake-useragent (>=2.2.0,<3.0.0)",
    "google-oauth2-tool (>=0.0.3,<0.0.4)",
    "scrapeops-python-requests (>=0.4.7,<0.5.0)",
    "setuptools (>=82.0.1,<83.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    {include = "src"},
]

[tool.poetry.scripts]
app = "app:main"

